Listen for mpv events and update state accordingly
At first I tried to use long polling to receive events in a browser.
The problem with that approach is that it's very easy to miss events.
So, now the persistent HTTP connection is used instead.
Through this connection I feed JSON objects from mpv to a browser.
This is achieved by not setting the Content-Length response header.

Windows-specific code is untested.